+++ lib/t/test_mp_opt_model.m
@@ -37,6 +37,10 @@ tests{end+1} = 't_qps_master';
 tests{end+1} = 't_miqps_master';
 tests{end+1} = 't_nlps_master';
 tests{end+1} = 't_opt_model';
+tests{end+1} = 't_om_solve_nleqs';
+tests{end+1} = 't_om_solve_qps';
+tests{end+1} = 't_om_solve_miqps';
+tests{end+1} = 't_om_solve_nlps';
 
 %% run the tests
 all_ok = t_run_tests( tests, verbose );